+++ include/io/file_type.hpp
@@ -1,0 +1,17 @@
+#pragma once
+
+#include "system/beg.hpp"
+
+#include <string>
+
+enum class FileType : uint8_t {
+    None,
+    File,
+    Directory,
+};
+
+namespace file_type {
+
+extern std::string getName(FileType id);
+
+}