+++ include/network/request/handlers/get_version.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "system/beg.hpp"
+
+#include "../handler.hpp"
+
+class GetVersionRH final : public RH {
+  public:
+    GetVersionRH() = default;
+
+    void run() override;
+};