+++ include/network/request/handlers/logout.hpp
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "system/beg.hpp"
+
+#include <boost/asio.hpp>
+#include "../handler.hpp"
+
+class LogoutRH : public RH {
+  private:
+    bool sendBeginResponse(boost::asio::ip::tcp::socket& socket);
+
+  public:
+    LogoutRH() = default;
+
+    void run() override;
+};