+++ include/users/fs_config.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "system/beg.hpp"
+
+namespace user_fs_config {
+
+static inline constexpr char PERSO_DIR_PREFIX[]{"perso"};
+
+}