+++ include/users/manager.hpp
@@ -1,0 +1,24 @@
+#pragma once
+
+#include "system/beg.hpp"
+
+#include <filesystem>
+#include <vector>
+#include "user.hpp"
+
+class UsersManager final {
+  private:
+    std::vector<User> users{};
+
+    UsersManager() = default;
+
+    bool checkUsersFolderExistence() const;
+
+  public:
+    static UsersManager& getInstance();
+
+    void load();
+    bool exists(const std::string& username);
+    User get(const std::string& username);
+    bool set(const User& user);
+};