+++ include/users/uif/values.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "system/beg.hpp"
+
+#include <string>
+
+struct UIFValues {
+    uint8_t role{};
+    std::string username{}, password{};
+};