+++ include/util/error_message.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "system/beg.hpp"
+
+#include <string>
+
+namespace util::error_message {
+
+extern std::string get(int errnum);
+
+} // namespace util::error_message