+++ include/util/thread.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "system/beg.hpp"
+
+#include <string>
+
+namespace util::thread {
+
+extern std::string getCurrentThreadIdAsString();
+
+}